+++ vaultier/accounts/api.py
@@ -66,16 +66,6 @@ class UserViewSet(AtomicTransactionMixin,
     permission_classes = (CanManageUserPermission,)
     model = User
 
-    def create(self, request, *args, **kwargs):
-        """
-        Check if is registration allowed or if is any registered user in here.
-        Otherwise raise 405 exception
-        """
-        if settings.VAULTIER.get('registration_allow') or not \
-                bool(User.objects.all().count()):
-            return super(UserViewSet, self).create(request, *args, **kwargs)
-        raise MethodNotAllowed(method='POST')
-
     def get_serializer_class(self):
         """
         Change serializer class to UserKeySerializer in list action