+++ vaultier/accounts/serializers.py
@@ -4,6 +4,7 @@ from rest_framework.request import Request
 from accounts.business.fields import RecoverTypeField, MemberStatusField
 from accounts.models import Token, User, LostKey, Member
 from nodes.models import Policy, Node
+from django.conf import settings
 
 
 class AuthSerializer(serializers.Serializer):
@@ -26,6 +27,40 @@ class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ['id', 'email', 'nickname', 'public_key']
+
+    def get_fields(self):
+        """
+        Adds invitation_hash field in create action only if registration is
+        disabled
+        """
+        request = self.context.get('request')
+        ret = super(UserSerializer, self).get_fields()
+
+        if not request:
+            return ret
+
+        if request.method == 'POST' and \
+                bool(User.objects.all().count() is not 0) and \
+                not settings.VAULTIER.get('registration_allow'):
+            ret['invitation_hash'] = serializers.CharField(write_only=True,
+                                                           required=True)
+        return ret
+
+    def validate_invitation_hash(self, attrs, source):
+        """
+        Validate invitation hash validity
+        """
+        hash = attrs[source]
+        try:
+            Member.objects.get(invitation_hash=hash)
+        except Member.DoesNotExist:
+            # first user can register
+            if bool(User.objects.all().count()):
+                # kick the rest
+                raise serializers.ValidationError("Invitation not found")
+
+        del attrs[source]
+        return attrs
 
     def restore_fields(self, data, files):
         if self.context.get('view').action != 'create':