+++ vaultier/vaultier/scripts/app/dal/model/Role.js
@@ -126,7 +126,7 @@ Vaultier.dal.model.Role = RL.Model.extend(
          */
         isRelatedToObject: function (object) {
             return this.get('relatedObjectType') === object.get('objectType') &&
-                this.get('relatedObject.id') === object.get('id');
+                this.get('relatedObject') === object.get('id');
         }
 
     });